% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_effect.R
\name{est_effect}
\alias{est_effect}
\title{Estimate Causal Effects}
\usage{
est_effect(object, Y, ...)
}
\arguments{
\item{object}{An object to calculate effects.}

\item{Y}{Numeric vector of observed outcomes.}

\item{...}{Additional arguments (ignored).}
}
\value{
Returns values of effects.
}
\description{
This is a generic function that estimates causal effects for different object types.
It automatically dispatches the appropriate method based on the class of `object`.
}
\details{
This function uses S3 method dispatching to call the appropriate method based on the object type.
It is designed for estimating causal effects in settings with continuous or binary outcomes.
For survival outcomes, users should apply appropriate survival analysis models, such as a
weighted Cox model or other time-to-event estimation methods. For `lbc_net` objects, see
\code{\link{est_effect.lbc_net}} for more details. Additional object types may be supported in the future.
}
