% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLBC.R
\name{getLBC.lsd}
\alias{getLBC.lsd}
\title{Extract Components from an lsd Object}
\usage{
\method{getLBC}{lsd}(object, name = "LSD_mean")
}
\arguments{
\item{object}{An object of class `"lsd"`, generated by \code{\link{lsd}}.}

\item{name}{A character vector specifying the name(s) of the component(s) to extract.
If `name = "ALL"`, a named list of all available components will be returned.
Available options include:
\describe{
  \item{`"LSD"`}{A matrix of LSD values for each covariate at each `ck`.}
  \item{`"LSD_mean"`}{The mean absolute LSD value across all covariates.}
  \item{`"LSD_max"`}{The maximum absolute LSD value.}
  \item{`"Z"`}{Covariate matrix used.}
  \item{`"Tr"`}{Treatment assignment vector.}
  \item{`"ck"`}{Kernel center values used.}
  \item{`"h"`}{Bandwidth values used.}
  \item{`"K"`}{Number of kernel points.}
  \item{`"rho"`}{Span values used.}
  \item{`"ate_flag"`}{ATE or ATT aim for estimate.}
  \item{`"kernel"`}{kernel used to define local neibourhood.}
}}
}
\value{
The requested component(s) from the `lbc_net` object. Default returns "fitted.values".
If multiple names are provided or if `name = "ALL"`, a named list of components is returned.
}
\description{
Retrieves specific components from an `lsd` object.
This function provides a structured way to access key results,
ensuring users retrieve the correct model components.
}
\examples{
\dontrun{
model <- lbc_net(data = data, formula = Tr ~ X1 + X2 + X3 + X4)
LSD.fit <- lsd(model)
getLBC(LSD.fit, "LSD")
getLBC(LSD.fit, c("LSD", "LSD_mean"))  # Extract multiple components
getLBC(LSD.fit, "ALL")  # Extract all components as a named list
}
}
