% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.lsd}
\alias{print.lsd}
\title{Print Basic Information of an lsd Object}
\usage{
\method{print}{lsd}(x, ...)
}
\arguments{
\item{x}{An object of class "lsd", generated by lsd().}

\item{...}{Additional arguments (ignored).}
}
\value{
Prints key training details and model parameters.
}
\description{
Provides a concise summary of an lsd object for local balance (LSD).
}
\examples{
\dontrun{
model <- lbc_net(data = data, formula = Tr ~ X1 + X2 + X3 + X4)
lsd_fit <- lsd(model)
print(fit)  # Displays a concise overview
}

}
\seealso{
\code{\link{summary}} and \code{\link{getLBC}}
}
