% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.lsd}
\alias{summary.lsd}
\title{Summary of an lsd Object}
\usage{
\method{summary}{lsd}(object, ...)
}
\arguments{
\item{object}{An object of class `"lsd"`, generated by `lsd()`.}

\item{...}{Additional arguments passed to the specific method.}
}
\value{
A structured summary of the `lbc_net` object, including:
\itemize{
  \item Average local balance over all cks for each covariate.
  \item Local balance summary: `lsd_max`, `lsd_mean` for local standardized mean difference calculated during training.
  \item Sample characteristics: `sample_size`, `num_covariates`, `treated_size`, `control_size`.
}
}
\description{
Provides a structured summary of an `lsd` object.
}
\details{
The function extracts key model components using \code{\link{getLBC}}.
}
\examples{
\dontrun{
model <- lbc_net(data = data, formula = Tr ~ X1 + X2 + X3 + X4)
lsd_result <- lsd(model)
summary(lsd_result)
}

}
\seealso{
\code{\link{getLBC}}
}
