% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLBC.R
\name{getLBC}
\alias{getLBC}
\title{Extract Components from an Object}
\usage{
getLBC(object, name)
}
\arguments{
\item{object}{An object from which to extract components.}

\item{name}{The name(s) of the component(s) to extract.}
}
\value{
The requested component(s), based on the specific method for the object's class.
}
\description{
This is a generic function used to extract components from different object types.
The function dispatches the appropriate method based on the class of `object`, ensuring
users can retrieve key model outputs from supported objects such as `"lbc_net"` and `lsd`.
}
\details{
This function uses S3 method dispatching to call the appropriate method based on the
object type. For example:
If `object` is of class `"lbc_net"` or `"lsd"`, see \code{\link{getLBC.lbc_net}} and \code{\link{getLBC.lsd}} for details.
Additional object types may be supported in the future.
}
\examples{
\dontrun{
fit <- lbc_net(data = my_data, formula = Tr ~ X1 + X2)
getLBC(fit, "fitted.values")  # Extract propensity scores
}

}
