% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_lbcnet.R
\name{setup_lbcnet}
\alias{setup_lbcnet}
\title{Setup Python Environment for LBCNet}
\usage{
setup_lbcnet(use_conda = FALSE, envname = "r-lbcnet")
}
\arguments{
\item{use_conda}{Logical. If `TRUE`, it attempts to use Conda (`r-lbcnet`) instead of virtualenv.
Default is `FALSE`, meaning virtualenv/system Python is preferred.}

\item{envname}{A character string specifying the name of the virtual environment or Conda environment
to use.  Default is `"r-lbcnet"`. One can use `virtualenv_list()` or `conda_list()` to check check the available Python environments in your system.}
}
\value{
This function configures the Python environment but does not return any value.
}
\description{
This function configures the Python environment for LBCNet, ensuring the correct
Python version and dependencies are loaded. It prioritizes virtual environments (`venv`) over
Conda unless explicitly requested.
}
\details{
- The function automatically detects the best available Python environment.
- If a user specifies `env`, it tries to activate that environment.
- It ensures required Python packages (`torch`, `numpy`, `pandas`, `tqdm`) are available using `py_require()`.
- If `RETICULATE_PYTHON` is set in `.Renviron`, it will respect that setting.
}
\examples{
\dontrun{
setup_lbcnet()  # Automatically configures the best available Python environment
setup_lbcnet(envname = "myenv")  # Uses a specific virtual environment
setup_lbcnet(use_conda = TRUE)  # Forces Conda if available
}

}
